# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SakuPlan API
  version: 0.1.0
  description: Core API for the SakuPlan personal budget planner.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/budgets/{id}/activate:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      tags: [Budgets]
      operationId: activateBudget
      responses:
        '200':
          description: Activated budget.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Budget' }
        '409': { $ref: '#/components/responses/Error' }
...
  /v1/bills:
    get:
      tags: [Bills]
      operationId: listBills
      parameters:
        - in: query
          name: include_inactive
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Recurring bills.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Bill' }
    post:
      tags: [Bills]
      operationId: createBill
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBillRequest' }
      responses:
        '201':
          description: Bill created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bill' }
  /v1/bills/{id}/occurrences:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Bills]
      operationId: markBillPaid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkBillPaidRequest' }
      responses:
        '201':
          description: Bill occurrence marked paid.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillOccurrence' }
  /v1/goals:
    get:
      tags: [Goals]
      operationId: listGoals
      parameters:
        - in: query
          name: include_inactive
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Saving goals.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Goal' }
    post:
      tags: [Goals]
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGoalRequest' }
      responses:
        '201':
          description: Goal created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Goal' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
...
    Error:
      description: Standard error envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    Money:
      type: integer
      format: int64
      description: Integer minor units. For IDR, one unit equals one rupiah.
      examples: [125000]
...
    BudgetAllocation:
      type: object
      required: [id, category_id, amount]
      properties:
        id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        amount: { $ref: '#/components/schemas/Money' }
    Budget:
      type: object
      required: [id, start_date, end_date, expected_income, savings_commitment, minimum_buffer, status, source, allocations]
      properties:
        id: { type: string, format: uuid }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        expected_income: { $ref: '#/components/schemas/Money' }
        savings_commitment: { $ref: '#/components/schemas/Money' }
        minimum_buffer: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [draft, active, closed] }
        source: { type: string, enum: [manual, rule_based, ai_assisted] }
        allocations:
          type: array
          items: { $ref: '#/components/schemas/BudgetAllocation' }
    CreateBillRequest:
      type: object
      required: [name, amount, due_day, frequency, category_id, account_id]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        amount: { $ref: '#/components/schemas/Money' }
        due_day: { type: integer, minimum: 1, maximum: 31 }
        frequency: { type: string, enum: [monthly] }
        category_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        reminder_days: { type: integer, minimum: 0, maximum: 90 }
    Bill:
      allOf:
        - $ref: '#/components/schemas/CreateBillRequest'
        - type: object
          required: [id, active, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            active: { type: boolean }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            last_paid_due_date:
              type: string
              format: date-time
              nullable: true
              description: due_date of the most recently paid bill_occurrence, or absent if none has ever been paid.
    MarkBillPaidRequest:
      type: object
      required: [due_date]
      properties:
        due_date: { type: string, format: date-time }
    BillOccurrence:
      type: object
      required: [id, bill_id, due_date, amount, transaction_id, created_at]
      properties:
        id: { type: string, format: uuid }
        bill_id: { type: string, format: uuid }
        due_date: { type: string, format: date-time }
        amount: { $ref: '#/components/schemas/Money' }
        transaction_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    CreateGoalRequest:
      type: object
      required: [name, target_amount, monthly_commitment, priority]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        target_amount: { $ref: '#/components/schemas/Money' }
        target_date: { type: string, format: date-time }
        monthly_commitment: { $ref: '#/components/schemas/Money' }
        priority: { type: integer }
    Goal:
      allOf:
        - $ref: '#/components/schemas/CreateGoalRequest'
        - type: object
          required: [id, status, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            status: { type: string, enum: [active, completed, archived] }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string, format: uuid }